requirements.txt
Pillow
albumentations
black
flake8
isort
lightning[extra]
loguru
mkdocs
omegaconf
opencv-python
polars
python-dotenv
pytorch_toolbelt
segmentation_models_pytorch
tensorboard
timm
torch
torchvision
tqdm
typer
-e .
